refactoring the material system towards  MaterialRenderer (previously known as Material ), MaterialInstance ( that takes the part of MaterialEntity's responsibilities as the material data storage ) and a MaterialEntity, which at the moment simply renders an entity with the specified MaterialInstance ( serves as a proxy between the scene tree and the material system ).

+++ trunk/Code/Engine/core-Renderer/RPPreviewNode.cpp
@@ -79,7 +79,7 @@ void RPPreviewNode::onUpdate( RenderingPipelineMechanism& host ) const
    RCBindPixelShader* comm = new ( renderer() ) RCBindPixelShader( *m_shader, renderer );
    {
       comm->setInt( "g_mode", (int)m_type );
-      comm->setInt( "g_farZ", activeCam.getFarClippingPlane() );
+      comm->setFloat( "g_farZ", activeCam.getFarClippingPlane() );
       comm->setTexture( "g_Texture", texture );
    }
